Extract field resolver
